Adjusted the code and resolved compilation errors

+++ dish/src/main/java/com/menu/dish/domain/model/DishInfo.java
@@ -2,12 +2,12 @@ package com.menu.dish.domain.model;
 
 public class DishInfo {
     private String name;
-    private int price;
+    private Double price;
     private String description;
     private String urlImage;
     private String category;
 
-    public DishInfo(String name, int price, String description, String urlImage, String category) {
+    public DishInfo(String name, Double price, String description, String urlImage, String category) {
         this.name = name;
         this.price = price;
         this.description = description;
@@ -18,8 +18,8 @@ public class DishInfo {
     public String getName() { return name; }
     public void setName(String name) { this.name = name; }
 
-    public int getPrice() { return price; }
-    public void setPrice(int price) { this.price = price; }
+    public Double getPrice() { return price; }
+    public void setPrice(Double price) { this.price = price; }
 
     public String getDescription() { return description; }
     public void setDescription(String description) { this.description = description; }
